refactor: Zodスキーマの型推論を活用し重複する型定義を削除
- src/types/database.tsを削除（Zodスキーマから型を自動生成するため）
- DatabaseErrorをuser-serviceで定義してエラー型を局所化
- CLAUDE.mdの型安全性ガイドラインを更新
